refactor: Update Rerender component import paths
Update the import paths for the Rerender component in the codebase to use lowercase file names. This change ensures consistency and follows best practices for file naming conventions.
